Add Godot project
Using Godot 4.4 stable with "Compatibility" setting and Git for
revisioning.

Change project settings to target 1080p which is a common resolution and
aspect ratio in modern PCs and consoles.

Also set project settings to warn if GDScript declarations aren't typed.
Static typing may not be a requirement for future learner contributions,
but it will be very helpful for the core team. As it improves
development with warnings and autocompletion, and gives better
performance.

Change icon to a custom PNG image, borrowed from initial concept art.

Add a main (empty) 2D scene.