Add basic Player scene
A top-down CharacterBody2D with an AnimatedSprite2D, and a capsule
collision shape for the feet in the ground to collide with walls.

Rename collision layers so number 1 is for the player, and 5 is for the
walls. A water tile will act as "wall" initially.

Add a new input action "running", and set it to keyboard Shift key. And
add WASD movement to the builtin ui_up, ui_down, ui_left, ui_right
actions; to have this other way of moving the player character (and
potentially navigating UI menus too) on PC.

Add moving script, with exports to control the walk and run speeds, the
stop and the moving acceleration steps.

Add an instance of the player to the main scene.

Add a camera to follow the player smoothly, with some world boundaries
to constrain the world a bit (2048x2048 pixels, or 32x32 tiles). And set
2x zoom to it.