Add basic tileset
Add a TileSet resource to be used in multiple tilemap layers.

For the water, add a physics layer so it acts as "wall". Still, add an
alternative water tile without physics that is needed for decoration.

Setup terrains so it's easier to draw tiles for grass, sand and bridges.

Add an animated foam for decoration.

Add multiple tilemap layers to the main scene, to create a little map.
Note: the reason it requires so many layers is because the Tiny Swords
asset pack is being used as provided. With a graphics editor, the tiles
could be simplified. But since this is supposed to be placeholder art,
invest more work in the Godot editor instead of trying to fix the
assets. At the expense of complexity and being not very performant.