Create 2 basic audio buses: Music & SFX
Having separate audio buses allows us to change the volume of all the audio that plays in that bus independently. It also allows adding effects to the buses, but we probably won't even use that. However, I think having at least 2 basic buses for Music and Audio, and have all AudioStreamPlayers use one of them (or a new bus inherited from one of them), would be a good start since that lets us have 2 sliders in the config to change the volume of the music and the sfx independently.