Added default scene for interact areas
This way, we have a scene that has the default configuration (collision layers and mask) of the interact area already setup.
I made it this way instead of by modifying the interact_area.gd because I'd like to let the contributors modify the collision layers of the nodes after they were created.

I think that this would not be needed anymore once we this pull request is merged into godot and default values for inherited properties can be overriden: https://github.com/godotengine/godot/pull/93787 -- https://github.com/godotengine/godot-proposals/issues/338