Added collectible items + HUD
Collectible items are interactable props that add new items to the
inventory.

Accessing the global GameState from the collectible item to not go through the player, because I kind of like that new interactables can be added without having to modify the player's code to be aware that they exist.

Added second list of items to game state. These items are only tracked for a quest, so that gets reset if a new quest starts.

Added logic to the InventoryItem class so we can have a texture for each type of inventory item.

Added HUD for inventory items within a quest. This HUD reads the GameState to display the amount of collected inventory items.