Apply MPL-2.0 license to original code
This is a soft copyleft license, which allows combining this project
with closed-source parts (such as, hypothetically, console ports of
Godot) while still requiring the source to this project itself to be
made available to those who receive an executable build of the game.

Add the license as a COPYING file. Explain in the README that this
applies to original code, and not to third-party assets and addons. (The
only addon at present is under the MIT license.)