Collectible item now has revealed state and can teleport
The revealed state is used when we want to place the collectible in a certain position in the editor, but we don't want it to be collectable until a condition is met. The collectible can also take you to a different level, I think we need a dedicated node for just 'teleporting' without an area or something like that, but for now I think this works.