Added collectibles to music puzzle and stealth level
In the music puzzle, the sheep + teleporter was replaced by an item that takes you to the next level

In the stealth level, the npc at the end was replaced by an item that takes you back to the world map.

At the moment of doing this, the combat challenge wasn't linked with the rest of the levels, so that will be done later.