ci: Report errors from `reuse lint`
When I added a job that runs `reuse lint`, I piped its output to `tee
$GITHUB_STEP_SUMMARY` so that the output shows up in a nicely-formatted
way in the workflow run summary.

Unfortunately, by default the exit status of a shell pipeline is the
exit status of the rightmost command. This means that even if `reuse
lint` fails, the pipeline will succeed, erroneously claiming that the
licensing information is correctly specified.

Use bash's pipefail option to cause the pipeline to fail in this case.