Player: Add fighting component
Copy the PlayerFighting node and subnodes from FightingPlayer to Player.

Add a state export to the player. It can be cozy (no fighting) or
fighting. Use this state to disable or enable processing of the new
PlayerFighting node and of the existing PlayerInteraction node.

Replace Player AnimationPlayer with the one in FightingPlayer. Which has
the same walk, idle animations plus new ones for fighting.

In the animation_player.gd script, use the Player state to behave
differently on process. The alternative solution is to add 2
AnimationPlayer nodes, one per state. But since the fighting also needs
walk and idle animations, this way seems more convenient, even when
adding an if/else condition that goes against the composition.