Add checkpoints
A checkpoint is just an Area2D + a SpawnPoint.

It's automatically activated when a player (collision layer = 1) steps on it.

There isn't any feedback to the player about the checkpoint activation, but I think this is fine.

On the technical side, I had to change the spawn_point code a bit to support spawn points that aren't owned by the current scene directly (in this case, checkpoint.tscn is the owner of the node). These spawnpoints DO NOT appear in the teleporter's inspector's dropdown, but can be set by code (as can be seen in checkpoint.gd)

Fixes https://github.com/endlessm/threadbare/issues/37