Adding a Player Controller
Since we need to be able to pause the inputs of a player, a Player Controller has been added as an intermediary between the Input singleton and the player scripts. This controller checks if the input has been paused, and if so, ignores them.

Player Controllers are also a good practice in general, since they allow to easily have player characters that are controlled by entities that are not input, like code for cinematics, or another player through multiplayer.