Add m6x11plus font
This is a pixel font that is usable at multiples of 18px. This is the
"plus" variant that "contains even more characters"; the non-plus
variant is usable at multiples of 16px. I chose the plus variant without
really investigating the difference on the basis that I have never
wanted fewer characters in my fonts.

Explicitly configure the font to not use subpixel positioning. This
happens automatically but with the default automatic setting we see this
message in the console:

    res://assets/fonts/m6x11plus.ttf: Pixel font detected, disabling subpixel positioning.

The license specified by the author is literally:

> free to use with attribution

Add this as a custom license.

https://managore.itch.io/m6x11