props/sign: Use m6x11 font
This matches the font now used for dialogue and interact labels.

For reasons I do not fully understand, the world map scenes (where these
signs are used) are scaled 2x by the camera. This affects the label
above the sign, but not the interact label (perhaps because the latter
is on a separate CanvasLayer?). As a result we need to set this sign's
font size to 18px to make it visually the same size as the interact
label's 36px font.