dialogue: Use Tiny Swords banner asset for main balloon
The main balloon is surrounded by a Panel, which the theme inherited
from Godot Dialogue Manager's default balloon scene previously specified
to be a StyleBoxFlat with a rounded border.

Replace that style with a 9-patch banner texture from Tiny Swords, with
the stretch mode set to tile (it's pixel art!) and its margins fudged a
little so that the whole box fits on the screen. There is a bit of a
weird interaction with the MarginContainer within - perhaps the margins
on that should be tweaked instead – but this looks a lot better already.

The buttons for dialogue choices are still in the default GDM style as
of this commit.

Helps https://github.com/endlessm/threadbare/issues/54