Tweak story quest starter dialogue
Previously, one of the choices given to the player was really long,
which looked bad. Simplify this to no/yes, with the extra flavour being
given in dialogue spoken by Storyweaver. Improve the surrounding
dialogue. Add a "maybe later" option which is functionally the same as
the "no" option but gives the illusion of broader choice, as well as
making it easy to test how it looks to have 3 options.