Dialogue: Style buttons
As previously, the "hover" and "normal" states are the same - the hover
state is not really used because the code automatically focuses the
button when the mouse enters it, and the focus state is drawn on top of
any other state.

Unselected buttons are blue. The focused item is yellow with a bright
yellow border.

While we don't currently use this GDM feature, it's possible to have a
dialogue choice shown but not available, in which case the button is
disabled. Style the disabled state consistently with the others in case
we need this later.
<https://github.com/nathanhoad/godot_dialogue_manager/blob/main/docs/Basic_Dialogue.md#responses>

Similarly, the "pressed" state is not used, but if we later introduced
toggle buttons somewhere else it might be nice to have that.

Helps https://github.com/endlessm/threadbare/issues/54