Fixed size label: Fix visibility
Fix regression when changing the interact label by the fixed size label
in 3aa25d56.

Setting visibility to the new label doesn't hide the label container
because it may have been reparented to the ScreenOverlay node.

So, add a generic object setter to the FixedSizeLabel and check if the
property being set is the visibility. If so, set the label container
visibility to the same value.