Added properties to the teleporters to be able to configure the transitions
In order to solve issue #69, the level designer should be able to choose which transitions are used during the teleporting sequence.

Three properties have been added to the teleporters. A boolean specifying if the teleporter should use transitions, and a couple of properties to specify an enter_transition and an exit_transition. These properties only appear when the "use_transition" boolean is true.

Fixes #67