Ink combat: Add walking to ink drinker
Export the period of time that the enemy throws ink. So it can be
adjusted per instantiated enemy, without having to edit the instance.
Call this "ink period".

Allow the enemy to walk randomly within a circle that has the initial
position as center. Export 3 attributes to define the walking:
- Walking time: how much of the ink period is used for walking. The
remaining part the enemy will be idle, with the attack happening in the
middle of the idle time. For example: if ink period is 5 seconds and
walking time is 3 seconds, then the enemy will be 1 second idle before
and 1 second idle after attacking.
- Walking range: The radius of the walking circle.
- Walking speed: The move speed when walking.

Add a walk animation to the animated sprite.

Add a State enum for walking, attacking and idle. To control the
behavior and playing the animations. The walking state is calculated
using the time left of the ink throwing timer. An alternative to this
would be to have multiple timers.

The walking circle is drawn in the editor canvas for enemies allowed to
walk.