Ink combat: Improve enemy walking area
When targetting the next position, skip a slice of the circle around the
enemy. And skip an inner circle. Effectively making the enemy target
points of a semi ring area.

Note that the enemy can still land in the skip areas, because it may not
have enough time to reach to the target position when walking, given its
current speed.

Adjust the debug circle to display the ring.