Integrate art for Houses, Tower, trees and Grass
Integrate the art of two different kind of houses, a tower, trees, and
grass to the different scenes.

Modify the Frays End scene to use the two kinds of houses.

The Grass texture had to be merged with the sand texture because
otherwise the tilemaps would break. Since the Sand was CC0, I believe we
can keep the CC-BY license in the derived work.