pre-commit: Exclude SVG files from whitespace checks
SVG files are, technically speaking, text files; but spiritually they
are image files that are generally edited with specialised tools rather
than with a text editor.

In particular, it is common for SVG files not to have a trailing
newline, and it seems cruel to force artists & developers to add one by
hand.

Exclude .svg files from trailing-whitespace and newline checks.