Player: Export sprite frames
So instances of player can have different look by changing the
SpriteFrames resouce.

If no frames are provided, it fallbacks to the default. The Story Weaver
is left as the default.

Since the frames are referenced in AnimationPlayer animations, the
SpriteFrames must have the following animation names with the following
number of frames:

- idle: 10 frames
- walk: 6 frames
- attack_01: 4 frames
- attack_02: 4 frames

Make the player script a tool so changing the sprite frames is visible
in the editor. This also makes the player mode disable the fighting node
in the editor.