Adding a lever that can be toggled and that can have targets
The lever can be on or off. And it can be toggled on or off, accordingly. When toggled, it emits a signal. It can also have "targets", which are a simpler interface than signals. Those targets should be of type "Toggleable". Although this is a "strong" contract, this is so the node picker in the editor only allows to pick those nodes that can be targets. Advanced users can use signals if they don't want to use or implement a Toggleable.