Add Baseline Sokoban Experience.
This Commit is bringing in a baseline working format for a Sokoban-style grid-based puzzle game. The structure of the sokoban mini-game currently is:

    Each tile or Piece2D is it's own scene with a list of groups and a layer id.
    Every piece lives as a child of the Board2D which manages adding, removing, getting and setting pieces by grid position
    There is a RuleEngine that takes in input from the DirectionalInput node, runs every rule, and processes movement/actions onto the board.

The rule engine does a few things:

    Provides directional movement to any piece under the player_groups
    Checks movement validity
    Runs every SokobanRule in rules to match & replace the gamestate before movement

A rule is a "match" pattern that will be matched against every single piece in the level, and a "replace" pattern that replaces the matched pieces. This allows for complex interactions to be formed from relatively simple rules. Every rule by default is applied after movement is started but before movement is finished.

The rule implementation includes:

    Direction: This dictates the direction of the pattern. (Is it oriented horizontally right, vertically up, horizontally left, vertically down)
    Context: While unused currently, this allows for the rule to be applied after the movement rather than before.
    Match Pattern: A list of RuleFragments to be matched against.
    Replace Pattern: A list of fragments to replace any matches with.

A rule fragment is a movement context and a group name. This uniquely identifies a type of piece and if it is staying still, moving, and what direction. This is used in a rule pattern as a series of pieces to match against.