MusicalRock: End interaction immediately, while sound plays
Previously, the rock interaction lasted as long as the sound effect.
This was annoying because the player was frozen on the spot for the
whole sound effect, rather than being able to move to the next note.

Don't await the end of the sound effect before ending the interaction.

Helps https://github.com/endlessm/threadbare/issues/112