Stella: Add skeleton intro scene
Following the lead of the MVP LoreQuest's closing scene, the "player" is
actually a plain AnimatedSprite2D, which is easier to animate than the
real Player scene.

The constellation asset is a placeholder, and the scene will probably
need to be rearranged when dialogue is moved to the left-hand side.

Fixes https://github.com/endlessm/threadbare/issues/191