Fill game: Replace color names with label + color
Decouple the label from the color in the fill barrels so it can be used
to create other kinds of games. The matching between projectile and
barrel now happens by label. Since a label is mandatory, use "???" as
default.

The color is now optional in both projectiles and barrels. And if
provided, will tint sprites as before to make a color-matching game.

Please note that if a color isn't provided, all projectiles will look
the same as of this commit. This is intended to change in upcoming
commits with an optional mapping of sprites per label.

Adjust the fill game logic accordingly.