Dialogue: Redesign balloon
This rebuilds the dialogue balloon scene to be tall rather than wide,
anchored to the top-left corner of the viewport, and with the speaker's
name pulled out to a ribbon at the top of the scroll. The
dialogue-choice buttons are moved into the scroll.

Compared to the previous implementation, derived from Godot Dialogue
Manager's example, I found it clearer to use PanelContainers rather than
using Panel and then carefully positioning the children. The theme is
updated accordingly.

Two styles of ribbon are defined for the speaker's name, BlueRibbon and
YellowRibbon. Only one is used in this implementation, but the intention
is that, in future, BlueRibbon will be used for the protagonist's name
and YellowRibbon for the names of NPCs.

Compared to the design, the balloon is always anchored at the top-left,
rather than being in the top-left for narration and bottom-left for
dialogue. Again this is deferred to a future change.

Helps https://github.com/endlessm/threadbare/issues/205