Fix wind shader calculations for large positions
There was a bug in the wind shader where large positions would all resolve to the same noise value, making the trees move at unison in those cases.

Instead of using the raw world position as a UV coordinate for the noise texture, we should use its fractional part so it remains between 0 and 1.