Redesign Fray's end
Remove connection with Song's Sanctuary.

The whole level is now delimited by cliffs at the top and the sides, and by a forest at the bottom.

There are 2 elders at the sides of the loom, one of them takes you to the lore quest and the other one is now a placeholder but that will later be changed to take you to the community story quests.

Related to https://github.com/endlessm/threadbare/issues/207