Add skeleton for ink combat rounds
In the first 2 ink combat rounds, replaced the collectible by a door + a teleporter.

Also, changed ink_combat.gd so it doesn't require a collectible and instead just emits a signal when it's complete.
Finally, made it possible for an ink_combat.gd to not have a starting dialogue. If there's no dialogue, the challenge will start 0.5 seconds after the level is loaded, if there's dialogue, it will start 0.5 seconds after the dialogue ended.

Fixes https://github.com/endlessm/threadbare/issues/241