Dialogue: Move character label inside panel
Previously the character label ribbon was a fixed size and anchored to the top of the panel.

Move the ribbon inside the panel. Adjust the panel margins to
accommodate this.

Left-justify the ribbon, rather than filling the horizontal space.
Change to a vanilla Label to make it easier to expand the label to the
text's natural size. The changing size of the ribbon draws the eye to a
change of speaker.

Fixes https://github.com/endlessm/threadbare/issues/158 (I hope!)