Add half-collision water to tileset
When using bridge tiles to build a south-facing pier, the last tile ends
halfway down its space. Add a new water alternative with a half-height
collision polygon for this case, which does not allow you to walk off
the end of the bridge.

Following commit d4c8aa459fe865812e4d3ee287c64eca5681563d, use the
non_walkable_floor physics layer: if this were used in an ink combat
level, we would not want the ink to collide with the water.