Change player attack from AnimatableBody to Area
Player's attack was a static body whose size grows when attacking. This caused unexpected behavior when trying to use the attack near another static body: if a projectile was between the player and a wall, the attack could push the projectile into the wall because it's pushing the rigidbody around every frame.

By having it by an area, we can make it so the attack only applies an impulse to the projectile on body entered, and if there's a wall it should just bounce off it.

Also, the way it's implemented now, we can make things be 'attackable' by having them in the colliding mask of the player's attack + implementing a hit_by method in those objects. This could help make the world a bit more interactive later.