splash: Remove debug line from scene
To make this animation, I manually drew the centre line as a Path2D and
width as a Curve of the Endless logo, tweaking both until it roughly
matches a reference logo. In order to do this in the editor, the scene
has a hidden Line2D which shows the resulting shape, along which the
stitching will be drawn.

The _update() function updates the Line2D's points field to the baked
points of the Path2D. However we have found that, on different GPUs
and/or OSes, the baked points are subtly different. And because this is
the default scene, which loads when you open Godot, the scene file ends
up getting edited in that case, meaning that contributors have to take
great care not to commit the spurious diff, and maintainers have to fix
PRs where this happens.

Remove the Line2D from the saved scene. Instead, when running in the
editor, construct it at runtime. Because its .owner is not set to the
scene, it does not get persisted into the saved scene.