Fix collectible that starts revealed
In 77834b0b the collectible animations were improved, and the sprite
starts with a transparent modulate color to avoid a flicker effect. But
the result of this is that if the collectible starts already revealed
(which is the case in the stealth game), the collectible is transparent
forever because the revealed animation doesn't play.

So in addition to set visibility when already revealed when ready, set
modulate color to white.