Loom: Don't jump to placeholder scene after incorporating threads
Previously, after completing a quest then interacting with the loom, the
player would be taken to a placeholder for what will one day be a
Sokoban puzzle.

Instead, add new lines of dialogue thanking StoryWeaver, and don't jump
to the placeholder. The lines are spoken by the "Elders" as a whole
because we can't easily tell which elder offered the player the quest
they just completed.