Direct StoryWeaver to Loom after completing quest
Previously, after completing a quest and returning to Fray's End, it was
possible to talk to an elder and start a new quest before taking the
threads you have just collected to the Eternal Loom.

Add new dialogue for both elders in this case. Update the misnamed
story_quest_starter.gd script (which in practice is actually the elder
script) to detect when the loom is ready for an offering and play the
alternative dialogue in this case.

Change the default _previous_look_at_side to UNSPECIFIED so that this is
preserved on the elder even though we have overridden the method that
would set it at the start of the dialogue.

Fixes https://github.com/endlessm/threadbare/issues/284

Co-authored-by: Andrea Pavón <andrea@endlessos.org>