Tweak dialogue introducing stealth quest
Currently this dialogue refers to InkDrinkers. But the guards in the
stealth quest are using a different asset, and we do not currently have
the necessary animations to replace them with InkDrinkers.

Refer to them as StoryVores in the dialogue that closes the preceding
scene instead.

Fixes https://github.com/endlessm/threadbare/issues/309