Create a stealth maze learner template scene
This simplified level is for learners to use to build their own
stealth maze.
It is placed into "template" folder within "StoryQuest" folder so it
sits with other learner templates.

Set name to parent node for the scene to "StealthTemplateLeveL"

Added cinematic node and tutorial text for learners within a new
stealth dialogue file.