Change default NPC & Talker sprite
Previously, the default sprite for an NPC or Talker was the guard with a
flaming torch.

We are using these guards as enemies in the stealth maze, so it doesn't
make sense for them also to be friendly NPCs you can talk to.

In particular, the input-tutorial NPC on the way into Fray's End should
not be a goblin guard.

Replace the default with a purple villager.