Change eternal loom animation
Made a new eternal loom animation that takes into account the size of the new threadspool assets, and plays a bit with their colors.

Also, since this animation changes the colors of the eternal loom, I think the default state of it could be to just have a base color instead of how it was working previously, when it was oscillating between different colors all the time.

Fixes https://github.com/endlessm/threadbare/issues/311