Cinematic: Add optional spawn_point_path
This will be used to place the player at a different position when they
first arrive at Fray's End (the path at the bottom-left) compared to
when they arrive back from a quest (in front of the loom).

It is a heavily stripped-back version of the same feature from
Teleporter, but just a bare string rather than implementing a tool
script to provide a drop-down. I did this because it is not trivial to
reuse the code from Teleporter, and we expect this only to be set in
the intro cinematic.

https://github.com/endlessm/threadbare/issues/293