Fray's End: Spawn at loom after completing a quest
After we redesigned this scene to have the player start outside Fray's
End proper, finishing a quest has regressed: the player should be placed
back near the loom in order to deposit the threads they have collected.

Move
the player's default location to near the loom. Add a spawn point at the
player's previous default location. Set the spawn_point_path property on
the intro cinematic.

I did it this way rather than the other way around (keep the player's
default position & place a spawn point in front of a loom) because this
way we only need to set the spawn point property in one scene that we
maintain, rather than requiring the property to be set correctly at the
close of every quest, which to me seems more error-prone.

Fixes https://github.com/endlessm/threadbare/issues/293