ci: Disable cancel-on-progress
Previously, all our workflows had this set on the basis that it is
wasted work to finish building the project or publishing the website if
another job is queued that will replace it.

However it causes the user that triggered the cancelled job to be
emailed about the job cancellation. There is no way I can find to
disable this email.

Disable this feature for now.