Add title scene with main menu
This screen shows a logo for the game, plus buttons to start the game or exit.

There are two hidden buttons for future options & credits scenes, which I added
here to test the appearance of the menu with more than two buttons.

The splash scene is adjusted to switch to this scene rather than the intro: the
title screen goes logically in between the two.

Fixes https://github.com/endlessm/threadbare/issues/33