Redesign Music Puzzle to make it clearer for players
Playtesting has shown that the music puzzle is hard to understand for
players. This was because of many reasons, and some of them have been
addressed in other commits, like the signs used to convey the melody.

One outstanding improvement was to make a whole level around the puzzle,
so players get more immersed and also have an opportunity to solve
easier puzzles before getting into the big one.

This commit changes the music_puzzle scene with many improvements:
- Extended the tilemap. It know features a level that emulates a lush
  forest.
- Split the puzzle: Instead of the musician giving you hints, they are
  spread in different areas.
- Added 3 new NPCs with additional dialogue
- The musician now directs you to explore the area and talk to the NPCs.
- Added smaller puzzles so the player can test their knowledge.
- Addded a script to randomize the trees.
- Packed the xylophone a bit so it can fit in one screen
- Added functionality in the puzzle and signs to allow for a puzzle to
  be partially solved from the get-go