Player: Handle input in _unhandled_input() not _process()
Previously, player input (directions, interacting, and repelling ink
blobs) was all done in the _process() method of Player, PlayerFighting
and PlayerInteraction. This meant that we had to take special steps to
avoid processing input during and at the end of dialogue.

Godot Dialogue Manager takes care to mark all input events as handled
during dialogue, and documents in its FAQ that player input should be
handled in _unhandled_input() rather than _process().

https://github.com/nathanhoad/godot_dialogue_manager/blob/main/docs/FAQ.md#how-do-i-stop-my-player-from-moving-while-dialogue-is-showing

Handle player input in _unhandled_input(). Two cases are simple:

- Player: add an input_vector field, update it from _unhandled_input(),
and use it in _process()
- PlayerFighting; rename _process() to _unhandled_input().

PlayerInteraction is slightly more complicated because we still
need to update the visibility of the interaction label in _process(): if
you keep an arrow key held down, the player's position (and potential
interaction) will change without any new unhandled input. So the
visibility of the label continues to be managed in _process(), while the
interact action itself is triggered in _unhandled_input().

https://github.com/endlessm/threadbare/issues/103

This change fixes an issue when a Cinematic node is used to trigger introductory
dialogue in a scene which also has an interactive Player, as we are doing in the
StoryQuest template's stealth template. Previously the player could be moved
during the introductory dialogue. Now it cannot.

Fixes https://github.com/endlessm/threadbare/issues/339