Bind Tab, not Esc, to skip dialogue
Previously Esc both paused the game (if no dialogue is playing) and skipped
dialogue (if it is). This was surprising behaviour for at least one player.

Bind Tab instead of Esc to skip dialogue. Heather said she tried Tab when trying
to figure out how to skip dialogue, and some web searches suggest that Tab is a
relatively common binding for this action in other games. (Another common choice
is to hold Ctrl to skip, but I personally do not expect modifiers to do anything
unless pressed together with another key, as is the case with Shift to run.)

Fixes https://github.com/endlessm/threadbare/issues/347