Bind P and Pause to pause
Currently, the only keyboard binding for the pause action is Escape.

Some keyboards have a key labelled Pause. Bind this to pause.

In addition, bind the P key to pause. Just as with WASD for motion, I bound the
physical location on a US English QWERTY keyboard rather than the key that
enters the letter P; unlike WASD, these are the same thing on the French AZERTY
layout.

Fixes https://github.com/endlessm/threadbare/issues/229