Projectile: Decouple hit player and enemies
Previously the projectile was able to either hit the player or the
enemies at the same time. The initial state was to hit the player, and
it switched to hit the enemy when the player pushes the projectile.

Split into two exports: "can hit player" and "can hit enemy" so the
projectile can be configured to do either both, none, or one of these.

Also move the particle changes when the player hits the ball to the
`hit_by()` method.