Dialogue: Use flat buttons for response menu
It has turned out that the previous button style makes it hard to tell which
choice is focus when there are only 2 choices: is it the blue button with a grey
border, or the dimmer yellow-brown button with a gold border?

Adjust the FlatButton style variation to show no background (actually a
fully-transparent modulation of the same texture) when unfocused, and the flat
sign texture when focused. Set the response buttons to use this.

Add a new FlatNextButton style, derived from FlatButton, that implements the
previous FlatButton style of having a visible background even when unfocused,
and set the "next >" button to use it. (This is because this button is not
actually focused: the dialogue balloon itself grabs focus, for reasons I don't
fully understand.)

I chose to define the styles this way around because I plan to use the same
style of button where only the focused one has a visible background for the main
menu.

Fixes https://github.com/endlessm/threadbare/issues/291