Fill game: Face enemies toward player
In two moments: when the enemy is ready, and when the enemy attacks.

The alternative is to do this every frame, but I think that's not
needed.

Note: The way the enemy is flipped horizontally is by scaling the whole
scene, not by using the AnimatedSprite2D flip_h property. This is
because the marker used to place the spawned projectile needs to be
flipped as well. Alternatively, we could store the initial marker
position as the left position, and a mirrored marked position as the
right position. Since scaling works and no regressions were found, I'd
say let's try this instead.

Fix https://github.com/endlessm/threadbare/issues/277