Add HUD to frays end, but have story progress visible only when all items were collected
The HUD's story progress visibility is updated on _ready() and also if the inventory changes. The point is that the story progress is visible if the eternal loom is possible to be activated. Right now there isn't any way to get new items without leaving fray's end, but if there were, this would keep working.

Fixes https://github.com/endlessm/threadbare/issues/329