Rework sokoban, add new levels, add new visuals. (#350)
A complete rework of the existing Sokoban. This commit does a few things
from the original sokoban implementation:

1. Simplify the user interface and replace resources with nodes.

This meant changing out the use of resources in the inspector for rules
and goals, to using node children and inheritance to create custom rules
and goals. This removed a lot the nested resources and makes it
easier/more intuitive how a sokoban is functioning.

This also meant removing some of the internal resources like
sokoban_fragment.gd and turning them into inner classes inside
pattern.gd.

Also the group resource was removed and replayed with ids and tags. Each
piece has an id that uniquely identifies it, and a list of tags that it
can share with other pieces. This then can be used in rules or goals for
complex behavior.

2. Add Query class to Board2D

Instead of an increasing list of parameters for each query function for
Board2D, I made a Query class that can be constructed and added to to
create a match against pieces. This should be expandable if needed in
the future.

3. Add go_to_scene on rule_engine.gd

When goals are completed, the rule engine will try to automatically load
the next scene. Currently this is set up to do three increasingly
difficult puzzles and then go back to fray's end.

4. Add Sokoban Dialogue to explain how to play

5. Add Undo and Reset abilities

The rule engine now remembers every step the player can take, and can
undo steps or reset the puzzle entirely with "z" and "r".

6. Automatically assign layers to pieces

Pieces do not need to be assigned a layer anymore, and will
automatically assume their layer based on which TileMapLayer it is a
part of. This means that pieces that are manually placed into the scene
will need to be placed as a child of a TileMapLayer to work properly.

7. Add new levels

There are currently 3 sets of levels, each with an A, B, and C section.
The A section starts with just one thread, then B adds a second, and
finally C has all three threads. After C, the level will go to fray's
end. Currently, the only level set 2 is set to work from the base game,
but that could be randomized in another pull request.

8. Add new visuals

There are now new visuals for every tile for the sokoban. On top of
this, there is a shader to darken every other tile to make it easier to
distinguish the different tiles. There is a new piece varient, the
"spool" for each color that acts like a wall.

Example of the new visuals!

![image](https://github.com/user-attachments/assets/e9e0fdea-1440-4a10-93a4-ab400b8ea357)