Indicate checkpoints visually
Previously, there was no visible indication that there are checkpoints in the
stealth level, or that you have reached one. In playtesting, some players
suggested that we should add checkpoints to the stealth level, not realising
that they were already present.

Add a sprite to the checkpoint scene. Hide it until the player enters the
checkpoint area.

Add a small collision shape to it, so that the player can't walk through it.
Deactivate that collision shape when the sprite is invisible.

Enable y-sorting on the parent node of all the checkpoints in the two stealth
levels currently in the source tree.

For now, a scarecrow is used to indicate the checkpoint, as a placeholder. I
have not added any kind of animation or sound effect when the checkpoint is
reached.

Fixes https://github.com/endlessm/threadbare/issues/240