Talker: Parametrise player name in default dialogue
Previously, the default dialogue hardcoded the player name to "StoryWeaver".
When a talker is added to a StoryQuest, the player's name will not be
"StoryWeaver".

While we expect that this default dialogue will not be used in-game, we can make
it work if it is. Use the `player_name` property of the `Player` scene, which is
passed into this dialogue as additional state by talker.gd when triggering the
dialogue.